refactor: 路由改为单向 URL→sessionView
去掉双向 Bridge：pathname 只调用 applyRoute；跳转显式用 Link/navigate
（New Session / Fork / Send / Inspect）。
